% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeByWeek.R
\name{SummarizeByWeek}
\alias{SummarizeByWeek}
\title{This function summarizes all columns by week and returns a dataframe with summarized data and week dates}
\usage{
SummarizeByWeek(df, summarizationMethod, flipNA)
}
\arguments{
\item{df}{Dataframe on which to operate. Must have column DateStamp}

\item{summarizationMethod}{String that indicates mean or sum}

\item{flipNA}{TRUE or FALSE to convert NAs to 0}
}
\description{
This function summarizes all columns by week and returns a dataframe with summarized data and week dates
}
\examples{
SummarizeByWeek(df, "mean", TRUE)
}
\keyword{summarize}
\keyword{week}
\keyword{weekly}

